Removal of duplicated code.
